change spec download method, make blobdownload util function
